+++ dot-env.template
@@ -95,5 +95,5 @@ OPENROUTER_API_KEY=...
 PIPER_BASE_URL=...
 
 # Smart turn
-LOCAL_SMART_TURN_MODEL_PATH=
-REMOTE_SMART_TURN_URL=
+LOCAL_SMART_TURN_MODEL_PATH=...
+FAL_SMART_TURN_API_KEY=...
